+++ ludwig/modules/embedding_modules.py
@@ -20,8 +20,9 @@ from torch import nn
 
 from ludwig.constants import TYPE
 from ludwig.modules.initializer_modules import get_initializer
+from ludwig.modules.ludwig_module import LudwigModule
 from ludwig.utils.data_utils import load_pretrained_embeddings
-from ludwig.utils.torch_utils import get_torch_device, LudwigModule
+from ludwig.utils.torch_utils import get_torch_device
 
 logger = logging.getLogger(__name__)
