+++ ludwig/modules/ludwig_module.py
@@ -1,0 +1,81 @@
+# Copyright (c) 2022 Predibase, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
+from abc import abstractmethod
+from functools import lru_cache
+
+import torch
+from torch.nn import Module, ModuleDict
+
+
+class LudwigModule(Module):
+    def __init__(self):
+        super().__init__()
+        self._losses = {}
+        self.register_buffer("device_tensor", torch.zeros(0))
+
+    @property
+    def device(self):
+        return self.device_tensor.device
+
+    def losses(self):
+        collected_losses = []
+        for loss in self._losses.values():
+            collected_losses.append(loss)
+
+        for child in self.children():
+            if isinstance(child, LudwigModule):
+                collected_losses.extend(child.losses())
+            elif isinstance(child, ModuleDict):
+                for c in child.values():
+                    if hasattr(c, "losses"):  # Some modules, i.e. SequenceReducers, don't have losses.
+                        collected_losses.extend(c.losses())
+            elif isinstance(child, Module):
+                pass
+            else:
+                raise ValueError
+
+        return collected_losses
+
+    def update_loss(self, key: str, loss: torch.Tensor):
+        """This should be called in the forward pass to add a custom loss term to the combined loss."""
+        self._losses[key] = loss
+
+    @property
+    def input_dtype(self):
+        return torch.float32
+
+    @property
+    @abstractmethod
+    def input_shape(self) -> torch.Size:
+        """Returns size of the input tensor without the batch dimension."""
+        raise NotImplementedError("Abstract class.")
+
+    @property
+    def output_shape(self) -> torch.Size:
+        """Returns size of the output tensor without the batch dimension."""
+        return self._compute_output_shape()
+
+    @lru_cache(maxsize=1)
+    def _compute_output_shape(self) -> torch.Size:
+        dummy_input = torch.rand(2, *self.input_shape, device=self.device)
+        output_tensor = self.forward(dummy_input.type(self.input_dtype))
+
+        if isinstance(output_tensor, torch.Tensor):
+            return output_tensor.size()[1:]
+        elif isinstance(output_tensor, dict) and "encoder_output" in output_tensor:
+            return output_tensor["encoder_output"].size()[1:]
+        else:
+            raise ValueError("Unknown output tensor type.")
